+++ src/views/EventVendorManagementSystem.vue
@@ -15,17 +15,15 @@
       </div>
       <div class="row justify-content-center mt-1">
         <div class="col-lg-11 mb-5 search-section-vendor">
-          <form @submit.prevent="searchVendor">
+          <form>
             <input
-              v-model="search.keyword"
               class="searchbox-input input"
               type="text"
               placeholder="What service are you searching for?"
             />
             <input
-              v-model="search.location"
               class="location-input input"
-              type="text" 
+              type="text"
               placeholder="Enter location"
             />
             <input
@@ -40,17 +38,14 @@
   </section>
   <!--Service Header Ends-->
 
-  <!--Advert Here-->
   <div class="container">
     <div class="row">
       <div class="col-lg-12 horizontal-advert mt-3">
-        <a :href="currentImg.ref_link" target="_blank">
+        <a href="#">
           <img
-            :src="currentImg.img_url"
+            src="@/assets/images/advert-banner3.jpg"
             ondragstart="return false;"
             alt="advert"
-            width="1300"
-            height="200"
           />
         </a>
       </div>
@@ -59,76 +54,385 @@
 
   <!--Service Event-Vendor-->
   <section class="event-vendor">
-        <div class="container service-vendor">
-            <div class="row">
-                <div class="col-lg-12 header mb-3">
-                    <h1>Available Jobs</h1>
+    <div class="container service-vendor">
+      <div class="row">
+        <div class="col-lg-12 header mb-4">
+          <h1>Popular Services</h1>
+        </div>
+        <div
+          v-if="error"
+          class="alert-danger alert alert-dismissible fade show"
+          role="alert"
+        >
+          {{ error }}
+          <button
+            type="button"
+            class="btn-close"
+            data-bs-dismiss="alert"
+            aria-label="Close"
+          ></button>
+        </div>
+        <div
+          v-if="message"
+          class="alert-success alert alert-dismissible fade show"
+          role="alert"
+        >
+          {{ message }}
+          <button
+            type="button"
+            class="btn-close"
+            data-bs-dismiss="alert"
+            aria-label="Close"
+          ></button>
+        </div>
+        <div class="col-lg-12">
+          <div class="row justify-content-center mainService-box">
+            <router-link
+              to="/category-vendor"
+              class="routers vendorService-box"
+            >
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-eventdecorator.png"
+                  ondragstart="return false;"
+                />
+                <h4>Event Decorators</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-photographer.png"
+                  ondragstart="return false;"
+                />
+                <h4>Photographers</h4>
+              </div>
+            </router-link>
+            <router-link
+              to="#"
+              class="routers vendorService-box"
+              ondragstart="return false;"
+            >
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-caterers.png"
+                  ondragstart="return false;"
+                />
+                <h4>Caterers</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-dj.png"
+                  ondragstart="return false;"
+                />
+                <h4>Disc Jockey (DJ's)</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-fashion.png"
+                  ondragstart="return false;"
+                />
+                <h4>Fashion Designers</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-eventplaner.png"
+                  ondragstart="return false;"
+                />
+                <h4>Event Planers</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-printing.png"
+                  ondragstart="return false;"
+                />
+                <h4>Invitations & Printing Services</h4>
+              </div>
+            </router-link>
+            <router-link to="#" class="routers vendorService-box">
+              <div>
+                <img
+                  src="@/assets/images/vendor-icon-makeup.png"
+                  ondragstart="return false;"
+                />
+                <h4>Make-Up Artist</h4>
+              </div>
+            </router-link>
+          </div>
+        </div>
+        <div class="col-lg-12 horizontal-advert mt-3">
+          <a href="#">
+            <img
+              src="@/assets/images/advert-banner3.jpg"
+              ondragstart="return false;"
+              alt="advert"
+            />
+          </a>
+        </div>
+      </div>
+    </div>
+
+    <div class="container service-vendor">
+      <div class="row">
+        <div class="col-lg-12 header mb-3">
+          <h1>Available Jobs</h1>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col-lg-9">
+          <div class="mainJob-box">
+            <div v-for="con in Content" :key="con.id" class="vendorJob-box">
+              <div class="card">
+                <div class="img-area">
+                  <img :src="con.event.image" />
                 </div>
+                <div class="card-body">
+                  <h1 title="Needed Service">{{ con.job_type }} Needed</h1>
+                  <div class="line-rule"></div>
+                  <span class="eventname" title="Event Type"
+                    ><i class="bi bi-stack"></i> {{ con.event.title }}</span
+                  >
+                  <span class="date" title="Bid Closing Date"
+                    ><i class="bi bi-calendar-week-fill"></i>
+                    {{ format_date(con.event.startdate) }}</span
+                  >
+                  <span class="location" title="Event Location"
+                    ><i class="bi bi-geo-alt-fill"></i> {{ con.location }}</span
+                  >
+                  <span class="price" title="Starting Price"
+                    ><i class="bi bi-cash"></i> Starts at
+                    <strong>&#8358;{{ con.budget }}</strong></span
+                  >
+                  <button
+                    type="button"
+                    data-bs-toggle="modal"
+                    data-bs-target="#exampleModal"
+                    v-on:click="getEventId(con.id)"
+                  >
+                    Bid
+                  </button>
+                </div>
+              </div>
             </div>
-            <div class="row">
-                <div v-for="con in Content" :key="con.id" class="vendorJob-box col-lg-3">
-                    <div class="card">
-                        <a :href="'/details-vendor/' + con.id">
-                            <div class="img-area">
-                                <img :src="con.event.image">
-                            </div>
-                        </a>
-                        <div class="card-body">
-                            <a :href="'/details-vendor/' + con.id" class="routers">
-                                <h1 title="Needed Service">{{ con.job_type }} Needed</h1>
-                            </a>
-                            <div class="line-rule"></div>
-                            <span class="eventname" title="Event Type"><i class="bi bi-stack"></i>
-                                {{ con.event.title }}</span>
-                            <span class="date" title="Bid Closing Date"><i class="bi bi-calendar-week-fill"></i>
-                                {{ format_date(con.event.startdate) }}</span>
-                            <!-- <span class="location" title="Event Location"><i class="bi bi-geo-alt-fill"></i>
-                                {{ con.location }}</span> -->
-                            <span class="price" title="Starting Price"><i class="bi bi-cash"></i> Starts at
-                                <strong>&#8358;{{ con.budget }}</strong></span>
-                            <button><a class="btnA" :href="'/details-vendor/' + con.id">View</a></button>
+            <!-- Modal -->
+            <div
+              class="modal fade vendorModalBox"
+              id="exampleModal"
+              tabindex="-1"
+              aria-labelledby="exampleModalLabel"
+              aria-hidden="true"
+            >
+              <div class="modal-dialog">
+                <div class="modal-content">
+                  <div class="modal-header">
+                    <h5 class="modal-title" id="exampleModalLabel">
+                      Bid For Job
+                    </h5>
+                    <button
+                      type="button"
+                      class="btn-close"
+                      data-bs-dismiss="modal"
+                      aria-label="Close"
+                    ></button>
+                  </div>
+                  <div class="modal-body">
+                    <form @submit.prevent="SubmitBid">
+                      <div v-if="loggedIn" class="row">
+                        <div
+                          v-if="error"
+                          class="alert-danger alert alert-dismissible fade show"
+                          role="alert"
+                        >
+                          {{ error }}
+                          <button
+                            type="button"
+                            class="btn-close"
+                            data-bs-dismiss="alert"
+                            aria-label="Close"
+                          ></button>
                         </div>
-                    </div>
+                        <div
+                          v-if="message"
+                          class="alert-success alert alert-dismissible fade show"
+                          role="alert"
+                        >
+                          {{ message }}
+                          <button
+                            type="button"
+                            class="btn-close"
+                            data-bs-dismiss="alert"
+                            aria-label="Close"
+                          ></button>
+                        </div>
+                        <div class="col-md-6 mb-3">
+                          <label>First Name</label>
+                          <input
+                            type="text"
+                            placeholder="Enter your name"
+                            class="input"
+                          />
+                        </div>
+                        <div class="col-md-6 mb-3">
+                          <label>Surname</label>
+                          <input
+                            type="text"
+                            placeholder="Enter your surname"
+                            class="input"
+                          />
+                        </div>
+                        <div class="col-md-6 mb-3">
+                          <label>Email Address</label>
+                          <input
+                            type="email"
+                            placeholder="Enter your email address"
+                            class="input"
+                          />
+                        </div>
+                        <div class="col-md-6 mb-3">
+                          <label>Telephone Number</label>
+                          <input
+                            type="tel"
+                            placeholder="Enter your phone nyumber"
+                            class="input"
+                          />
+                        </div>
+                        <div class="col-md-12 mb-3">
+                          <label>Place your bid (Naira &#8358;)</label>
+                          <input
+                            v-model="bid.price"
+                            type="number"
+                            placeholder="Enter bid amount"
+                            class="input"
+                          />
+                        </div>
+                        <div class="col-md-12 mb-3">
+                          <label>Add more info</label>
+                          <div class="textarea" contenteditable></div>
+                        </div>
+                        <div class="col-lg-12 mt-4">
+                          <label for="event image">Vendor Service Image</label>
+                          <input
+                            class="input"
+                            type="file"
+                            ref="file"
+                            accept=".jpg, .jpeg, .png, .jfif"
+                            v-on:change="handleFileUpload()"
+                            required
+                          />
+                          <small class="text-danger font-weight-bold"
+                            >(only .jpg, .jpeg, .png, .jfif or .webp
+                            format)</small
+                          >
+                        </div>
+                        <div class="col-md-12 mb-3">
+                          <button type="submit" :disabled="loading">
+                            Submit Bid<span
+                              v-show="loading"
+                              class="spinner-border spinner-border-sm"
+                            ></span>
+                          </button>
+                        </div>
+                      </div>
+                      <div v-else class="row">
+                        <div class="col-md-12 mb-3">
+                          <h3>You have to login to bid for jobs</h3>
+                        </div>
+
+                        <div class="event-form-content col-md-12 mb-3">
+                          <router-link to="/login" class="routers"
+                            ><a class="btn-view">login</a></router-link
+                          >
+                        </div>
+                      </div>
+                    </form>
+                  </div>
                 </div>
-                <!-- Modal -->
-
+              </div>
             </div>
-
-
-            <!-- <div class="col-md-12 text-center mt-5 paginationDivArea">
-                <nav aria-label="Page navigation example">
-                    <ul class="pagination justify-content-center">
-                        <li class="page-item">
-                            <a class="page-link" href="#" aria-label="Previous">
-                                <span aria-hidden="true">&laquo;</span>
-                            </a>
-                        </li>
-                        <li class="page-item"><a class="page-link" href="#">1</a></li>
-                        <li class="page-item"><a class="page-link" href="#">2</a></li>
-                        <li class="page-item"><a class="page-link" href="#">3</a></li>
-                        <li class="page-item">
-                            <a class="page-link" href="#" aria-label="Next">
-                                <span aria-hidden="true">&raquo;</span>
-                            </a>
-                        </li>
-                    </ul>
-                </nav>
-            </div> -->
-
-        </div>
-
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-12 horizontal-advert mt-3">
-                    <a :href="currentImg.ref_link" target="_blank">
-                        <img :src="currentImg.img_url" ondragstart="return false;" alt="advert" width="1300"
-                            height="200">
-                    </a>
+          </div>
+          <div class="mainJob-box mt-5">
+            <div v-for="con in Content" :key="con.id" class="vendorJob-box">
+              <div class="card">
+                <div class="img-area">
+                  <img :src="con.event.image" />
                 </div>
+                <div class="card-body">
+                  <h1 title="Needed Service">{{ con.job_type }} Needed</h1>
+                  <div class="line-rule"></div>
+                  <span class="eventname" title="Event Type"
+                    ><i class="bi bi-stack"></i> {{ con.event.title }}</span
+                  >
+                  <span class="date" title="Bid Closing Date"
+                    ><i class="bi bi-calendar-week-fill"></i>
+                    {{ format_date(con.event.startdate) }}</span
+                  >
+                  <span class="location" title="Event Location"
+                    ><i class="bi bi-geo-alt-fill"></i> {{ con.location }}</span
+                  >
+                  <span class="price" title="Starting Price"
+                    ><i class="bi bi-cash"></i> Starts at
+                    <strong>&#8358;{{ con.budget }}</strong></span
+                  >
+                  <button
+                    type="button"
+                    data-bs-toggle="modal"
+                    data-bs-target="#exampleModal"
+                  >
+                    Bid
+                  </button>
+                </div>
+              </div>
             </div>
+          </div>
+          <div class="col-md-12 text-center mt-5">
+            <a href="/view-more-vendor" class="btn-all-service">View More</a>
+          </div>
         </div>
+        <div class="col-lg-3">
+          <div class="advert-area">
+            <a href="#">
+              <img
+                src="@/assets/images/advert-banner1.jpg"
+                ondragstart="return false;"
+                alt="advert"
+              />
+            </a>
+          </div>
+          <div class="advert-area">
+            <a href="#">
+              <img
+                src="@/assets/images/advert-banner1.jpg"
+                ondragstart="return false;"
+                alt="advert"
+              />
+            </a>
+          </div>
+        </div>
+      </div>
+    </div>
 
-    </section>
+    <div class="container">
+      <div class="row">
+        <div class="col-lg-12 horizontal-advert mt-3">
+          <a href="#">
+            <img
+              src="@/assets/images/advert-banner3.jpg"
+              ondragstart="return false;"
+              alt="advert"
+            />
+          </a>
+        </div>
+      </div>
+    </div>
+  </section>
   <!--Service Event-Vendor Ends-->
 
   <elfrique-footer />
@@ -140,7 +444,7 @@ import Newsletter from "./elfrique-newsletter.vue";
 import Footer from "./elfrique-footer.vue";
 import VendorService from "../service/vendor.service";
 import moment from "moment";
-import $ from "jquery";
+
 export default {
   name: "Elfrique",
   components: {
@@ -151,7 +455,6 @@ export default {
   data() {
     return {
       Content: "",
-      adsContent: "",
       eventId: "",
       file: "",
       bid: {
@@ -161,68 +464,23 @@ export default {
       message: "",
       error: "",
       loading: false,
-      currentImage: {
-        img_url: "",
-        link: "",
-      },
-      timer: null,
-      currentIndex: 0,
-      search: {
-        location: "",
-        keyword: "",
-      },
     };
   },
   computed: {
     loggedIn() {
       return this.$store.state.auth.status.loggedIn;
     },
-    currentImg: function () {
-      return this.adsContent[
-        Math.abs(
-          this.currentIndex + Math.floor(Math.random() * this.adsContent.length)
-        ) % this.adsContent.length
-      ];
-    },
-    currentImg2: function () {
-      return this.adsContent[
-        Math.abs(
-          this.currentIndex + Math.floor(Math.random() * this.adsContent.length)
-        ) % this.adsContent.length
-      ];
-    },
   },
   created() {
-    VendorService.getAllJobs().then((response) => {
-      this.Content = response.data;
-    });
-
-    VendorService.getAllAds().then((response) => {
-      this.adsContent = response.data.data;
-    });
+    VendorService.getAllJobs()
+      .then((resp) => resp.json())
+      .then((response) => {
+        this.Content = response.data;
+        console.log(this.Content);
+      });
   },
 
   methods: {
-    searchVendor() {
-      this.$router.push({
-        name: "SearchEventVendor",
-        params: {
-          location: this.search.location,
-        },
-        query: {
-          keyword: this.search.keyword,
-        },
-      });
-    },
-
-    startSlide: function () {
-      this.timer = setInterval(this.next, 6000);
-    },
-
-    next: function () {
-      this.currentIndex += 1;
-    },
-
     getEventId(id) {
       this.eventId = id;
     },
@@ -261,20 +519,9 @@ export default {
       this.file = this.$refs.file.files[0];
     },
   },
-  mounted() {
-    this.startSlide();
-    window.scrollTo(0, 0);
 
-    $(document).ready(function () {
-      $(".vendorService-box").slice(0, 4).show();
-      $("#loadMore").on("click", function (e) {
-        e.preventDefault();
-        $(".vendorService-box:hidden").slice(0, 4).slideDown();
-        if ($(".vendorService-box:hidden").length == 0) {
-          $("#loadMore").text("No Service").addClass("noContent");
-        }
-      });
-    });
+  mounted() {
+    window.scrollTo(0, 0);
   },
 };
 </script>
